refactor(core): Refactor StreamWriter
