+++ src/main/java/org/sustain/handlers/StreamWriter.java
@@ -1,50 +1,55 @@
-package org.sustain.querier;
+package org.sustain.handlers;
 
 import io.grpc.stub.StreamObserver;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import java.util.ArrayList;
-import java.util.Iterator;
+import java.util.List;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.locks.ReentrantLock;
 
 public abstract class StreamWriter<T, R> {
+
+    private static final Logger log = LogManager.getLogger(StreamWriter.class);
+
     protected AtomicLong activeCount;
     protected LinkedBlockingQueue<T> queue;
-    protected ArrayList<Thread> threads;
+    protected List<Thread> threads;
 
     public StreamWriter(StreamObserver<R> responseObserver,
-            int threadCount) {
+                        int threadCount) {
         this.activeCount = new AtomicLong(0);
-        this.queue = new LinkedBlockingQueue();
-        this.threads = new ArrayList();
+        this.queue = new LinkedBlockingQueue<T>();
+        this.threads = new ArrayList<Thread>();
 
         // initialize threads
         ReentrantLock lock = new ReentrantLock();
         for (int i = 0; i < threadCount; i++) {
-            Thread thread = new Thread(new Runnable() {
-                @Override
-                public void run() {
-                    try {
-                        while (true) {
-                            // retrieve next document
-                            T t = queue.take();
+            Thread thread = new Thread(() -> {
 
-                            // initialize response
-                            R r = convert(t);
+                try {
+                    while (true) {
+                        // retrieve next document
+                        T t = queue.take();
 
-                            // send response
-                            lock.lock();
-                            try {
-                                responseObserver.onNext(r);
-                            } finally {
-                                lock.unlock();
-                            }
+                        // initialize response
+                        R r = convert(t);
 
-                            // decrement active count
-                            activeCount.decrementAndGet();
+                        // send response
+                        lock.lock();
+                        try {
+                            responseObserver.onNext(r);
+                        } finally {
+                            lock.unlock();
                         }
-                    } catch (InterruptedException e) {}
+
+                        // decrement active count
+                        activeCount.decrementAndGet();
+                    }
+                } catch (InterruptedException e) {
+                    log.info("Caught InterruptedException: " + e.getMessage());
                 }
             });
 
@@ -55,7 +60,7 @@ public abstract class StreamWriter<T, R> {
     public void add(T t) throws Exception {
         // add datum to processing queue and increment active count
         this.queue.add(t);
-		this.activeCount.incrementAndGet();
+        this.activeCount.incrementAndGet();
     }
 
     public abstract R convert(T t);
@@ -74,7 +79,7 @@ public abstract class StreamWriter<T, R> {
                 Thread.sleep(50);
             }
         }
-        
+
         // stop all threads
         for (Thread thread : this.threads) {
             thread.interrupt();
